work on AI state mechine
enemies now return to their patrols if they lose sight of the player. an
enemy will go to the last spot it saw the player then give up, later I
will added a search state and the enemy will search for the player and
if it can't find them it will give up and return to patrolling.