Fixed Camera Bug
The camera in the editor was zoomed out because the particle system was
set. I made it so it doesn't set the particle system in the Constructor
but in Begin Play this brought the camera back in and the particle
system still works at run time.

Warning this means the particle system can not be set in the editor
anymore. It has to be gone in code.

Note make the particle system editable in the editor again later.