just started
screwed around with swinging door learned something but no progress may
need to remove work or just start over.

got ray tracing to work, its displays the name of the object it hits.
need to move it to Tick function and not on a key prass.

note I needed to add a camera to get the ray trace to work but this
replaced the default camera. I need to fix the carmera now to get it
back to where it was.