Hiding spots and bill boards
add hiding spots so the player can hide from the enemies in them, it may
not be the best way to do it.

it is difficult to select things in the editor when they are see through
so I add bill board to some of the actors that need them, now I can just
click on the bill board.

add planks.

worked on shaders. made materials to help the player see in dim places.