Bug fixes
added glass material and tested how it looked in chamber02

fixed terrain in chamber07 so player can no longer get stuck

fixed the water not being set properly when chamber11 was load from
another level. add basic stuff in level blueprint to get it to work

removed power grid code so the unreal projected wouldn't brake when I'm
working on it. I'll work on it in a visual studio project and add it
back in at points of a working version