Up Date Power Gird And InteractAbles
remaded / renamed Interactables

made power gird classes and gave them all blue widget so everyone can
see them in the editor

made a new lightswitch = IA_Switch it is an interactable

made a Gate actor it can switch between Or, And, Not. but all of the
power girds now have a 0.5 second delay becasue Gate has to be set up
start so all new power grid object need that too

at al last I have defeated my arch nemesis The Door, it can open and
close. need to see if it works with the nav mesh though

made a demo for burnable objects, it needs to be reviewed

made ActorRespone just because, it respones when it falls off the world
but is not an interactable

added IntactWithPlayer to interactable and updated the player class to
use it, still need to make InteractWithObject